requirements.txt
SQLAlchemy==2.0.32
alembic==1.13.2
Faker==26.0.0
python-dotenv==1.0.1
